Move routing logic into <Explorer /> from <Sidebar />
